% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contacts_V1_jittered_100m}
\alias{contacts_V1_jittered_100m}
\title{FluScape contact survey data (jittered 100m)}
\format{
A data frame with 4710 rows and 14 variables; see
  \code{\link{contacts_fluscape_V1}} for column definitions.
}
\usage{
contacts_V1_jittered_100m
}
\description{
Same records as \code{\link{contacts_fluscape_V1}}, but with household
coordinates randomly jittered by up to 100m to allow public release
without disclosing exact household locations.
}
\keyword{datasets}
